requirements.txt
matplotlib==3.8.3
wandb==0.17.0
numpy==1.26.0
pandas==2.2.1

torch==2.0.1+cu118
torchvision==0.15.2+cu118
pillow==10.2.0